+++ system/src/Grav/Common/Inflector.php
@@ -161,9 +161,15 @@ class Inflector
      */
     public static function titleize($word, $uppercase = '')
     {
-        $uppercase = $uppercase === 'first' ? 'ucfirst' : 'ucwords';
+        $humanize_underscorize = static::humanize(static::underscorize($word));
 
-        return $uppercase(static::humanize(static::underscorize($word)));
+        if ($uppercase === 'first') {
+            $firstLetter = mb_strtoupper(mb_substr($humanize_underscorize, 0, 1, "UTF-8"), "UTF-8");
+            return $firstLetter . mb_substr($humanize_underscorize, 1, mb_strlen($humanize_underscorize, "UTF-8"), "UTF-8");
+        } else {
+            return mb_convert_case($humanize_underscorize, MB_CASE_TITLE, 'UTF-8');
+        }
+
     }
 
     /**
@@ -180,7 +186,7 @@ class Inflector
      */
     public static function camelize($word)
     {
-        return str_replace(' ', '', ucwords(preg_replace('/[^A-Z^a-z^0-9]+/', ' ', $word)));
+        return str_replace(' ', '', ucwords(preg_replace('/[^\p{L}^0-9]+/', ' ', $word)));
     }
 
     /**
@@ -198,7 +204,7 @@ class Inflector
     {
         $regex1 = preg_replace('/([A-Z]+)([A-Z][a-z])/', '\1_\2', $word);
         $regex2 = preg_replace('/([a-zd])([A-Z])/', '\1_\2', $regex1);
-        $regex3 = preg_replace('/[^A-Z^a-z^0-9]+/', '_', $regex2);
+        $regex3 = preg_replace('/[^\p{L}^0-9]+/u', '_', $regex2);
 
         return strtolower($regex3);
     }
@@ -219,7 +225,7 @@ class Inflector
         $regex1 = preg_replace('/([A-Z]+)([A-Z][a-z])/', '\1-\2', $word);
         $regex2 = preg_replace('/([a-z])([A-Z])/', '\1-\2', $regex1);
         $regex3 = preg_replace('/([0-9])([A-Z])/', '\1-\2', $regex2);
-        $regex4 = preg_replace('/[^A-Z^a-z^0-9]+/', '-', $regex3);
+        $regex4 = preg_replace('/[^\p{L}^0-9]+/', '-', $regex3);
 
         $regex4 = trim($regex4, '-');